separates tree components from demo stuff properly
Webpack is such a bitch. Checkout the change on line 6 in webpack.config.js. I had to wrap that in an array simply because it was used in the other entry point's build. What‽! Who makes a build tool like that! Solution was found here: https://github.com/webpack/webpack/issues/300

Gah! F**king webpack!!!